Implement the Ratio method
The ratio method is in total slower than the polar method, but only produces one output. This can be very nice for libraries as the normal distribution function doesn't require any state and is fast and accurate.
I couldn't get the AVX code to compile, so I couldn't update the timing tables.